move apiClient init in a composable
