+++ .eslintrc-auto-import.json
@@ -324,6 +324,11 @@
     "storeRouteMapping": true,
     "useClipboardItems": true,
     "markdown": true,
-    "usePerms": true
+    "usePerms": true,
+    "api": true,
+    "apiClient": true,
+    "jwt": true,
+    "tryRequest": true,
+    "useApiClient": true
   }
 }
