+++ auto-imports.d.ts
@@ -9,6 +9,8 @@ declare global {
   const acceptHMRUpdate: typeof import('pinia')['acceptHMRUpdate']
   const afterAll: typeof import('vitest')['afterAll']
   const afterEach: typeof import('vitest')['afterEach']
+  const api: typeof import('./src/composables/useApiClient')['api']
+  const apiClient: typeof import('./src/composables/useApiClient')['apiClient']
   const assert: typeof import('vitest')['assert']
   const asyncComputed: typeof import('@vueuse/core')['asyncComputed']
   const autoResetRef: typeof import('@vueuse/core')['autoResetRef']
@@ -63,6 +65,7 @@ declare global {
   const isRef: typeof import('vue')['isRef']
   const isString: typeof import('./src/utils/typeGuards')['isString']
   const it: typeof import('vitest')['it']
+  const jwt: typeof import('./src/composables/useApiClient')['jwt']
   const makeDestructurable: typeof import('@vueuse/core')['makeDestructurable']
   const mapActions: typeof import('pinia')['mapActions']
   const mapGetters: typeof import('pinia')['mapGetters']
@@ -135,12 +138,14 @@ declare global {
   const tryOnMounted: typeof import('@vueuse/core')['tryOnMounted']
   const tryOnScopeDispose: typeof import('@vueuse/core')['tryOnScopeDispose']
   const tryOnUnmounted: typeof import('@vueuse/core')['tryOnUnmounted']
+  const tryRequest: typeof import('./src/composables/useApiClient')['tryRequest']
   const unref: typeof import('vue')['unref']
   const unrefElement: typeof import('@vueuse/core')['unrefElement']
   const until: typeof import('@vueuse/core')['until']
   const uploadContent: typeof import('./src/composables/upload')['uploadContent']
   const useActiveElement: typeof import('@vueuse/core')['useActiveElement']
   const useAnimate: typeof import('@vueuse/core')['useAnimate']
+  const useApiClient: typeof import('./src/composables/useApiClient')['useApiClient']
   const useArrayDifference: typeof import('@vueuse/core')['useArrayDifference']
   const useArrayEvery: typeof import('@vueuse/core')['useArrayEvery']
   const useArrayFilter: typeof import('@vueuse/core')['useArrayFilter']
